requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch
Pillow
flask
requests
numpy
colpali-engine
flask-cors
